% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_to_tmb.R
\name{typical_callable_bases}
\alias{typical_callable_bases}
\title{Typical number of callable bases for common assays.}
\usage{
typical_callable_bases()
}
\value{
a list where names are sequencing strategy and values are the number of callable bases.
}
\description{
It is best-practice to identify the exact number of \strong{callable_bases}
based on your sequencing strategy, bioinformatics pipeline and breadth & depth of coverage (extracted from sample BAMs).
This function describes some sensible defaults for common sequencing strategies that should only be used if more comprehensive methods of estimation are not possible.
}
\examples{
typical_callable_bases()
}
